add:工作流设计--3%
(cherry picked from commit bdb370d)

+++ src/main/java/com/hzl/hadoop/workflow/entity/ApproveHistoryApproverEntity.java
@@ -48,7 +48,7 @@ public class ApproveHistoryApproverEntity implements Serializable {
 	/**
 	 * 审批动作，1同意，2拒绝，3跳过，4转交
 	 */
-	private Long approveAction;
+	private Integer approveAction;
 	/**
 	 * 租户id
 	 */
